don't retry failed requests
retryOnConnectionFailure is still set to true, but this was previously
silently retrying failures like timeouts.

If this leads to issues because the catch was catching other exceptions
in the construction of RNFetchBlobDefaultResp or RNFetchBlobFileResp, we
can restore the catch in a different place and still avoid making the
request twice.

Note that this stops setting timeout, because that's not actually used
for unsuccessful requests (see the onFailure handler).