rm timeout special handling
We weren't passing the response info value, so no need to set it. The
localized message also didn't seem valuable to provide - since these
messagse are internal in general. If clients should be able to
distinguish timeout failures in the fturue, we can add response info
metadata back and actually start surfacing that.